refactor(format): split plain object display helpers
